+++ tests/integration-tests/clusters_factory.py
@@ -18,6 +18,7 @@ import boto3
 import yaml
 from retrying import retry
 from utils import (
+    dict_add_nested_key,
     retrieve_cfn_outputs,
     retrieve_cfn_parameters,
     retrieve_cfn_resources,
@@ -46,14 +47,14 @@ class Cluster:
         attrs = ", ".join(["{key}={value}".format(key=key, value=repr(value)) for key, value in self.__dict__.items()])
         return "{class_name}({attrs})".format(class_name=self.__class__.__name__, attrs=attrs)
 
-    def update(self, force=True):
+    def update(self, config_file, force=True):
         """
         Update a cluster with an already updated config.
         :param force: if to use --force flag when update
         """
         # update the cluster
-        logging.info("Updating cluster {0} with config {1}".format(self.name, self.config_file))
-        command = ["pcluster", "update", "--config", self.config_file]
+        logging.info("Updating cluster {0} with config {1}".format(self.name, config_file))
+        command = ["pcluster", "update", "--config", config_file]
         if force:
             command.append("--force")
         command.append(self.name)
@@ -63,20 +64,39 @@ class Cluster:
             logging.error(error)
             raise Exception(error)
         logging.info("Cluster {0} updated successfully".format(self.name))
+        # Only update config file attribute if update is successful
+        self.config_file = config_file
+        with open(self.config_file) as conf_file:
+            self.config = yaml.load(conf_file, Loader=yaml.SafeLoader)
 
         # reset cached properties
         self._reset_cached_properties()
 
         return result
 
-    def delete(self, keep_logs=False):
+    def delete(self, delete_logs=False):
         """Delete this cluster."""
         if self.has_been_deleted:
             return
         cmd_args = ["pcluster", "delete", self.name]
-        if keep_logs:
+        if delete_logs:
+            logging.warning("Updating stack %s to delete CloudWatch logs on stack deletion.", self.name)
+            try:
+                dict_add_nested_key(
+                    self.config,
+                    False,
+                    ("Monitoring", "Logs", "CloudWatch", "RetainOnDelete"),
+                )
+                with open(self.config_file, "w") as conf_file:
+                    yaml.dump(self.config, conf_file)
+                self.update(self.config_file, force=True)
+            except subprocess.CalledProcessError as e:
+                logging.error(
+                    "Failed updating cluster to delete log with with error:\n%s\nand output:\n%s", e.stderr, e.stdout
+                )
+                raise
+        else:
             logging.warning("CloudWatch logs for cluster %s are preserved due to failure.", self.name)
-            cmd_args.append("--keep-logs")
         try:
             result = run_command(cmd_args, log_error=False)
             if "DELETE_FAILED" in result.stdout:
@@ -259,29 +279,32 @@ class ClustersFactory:
         logging.info("Sleeping for 30 seconds in case cluster is not ready yet")
         time.sleep(30)
 
-    @retry(stop_max_attempt_number=5, wait_fixed=5000, retry_on_exception=retry_if_subprocess_error)
-    def destroy_cluster(self, name, keep_logs=False):
+    def destroy_cluster(self, name, delete_logs):
         """Destroy a created cluster."""
         logging.info("Destroying cluster {0}".format(name))
         if name in self.__created_clusters:
-            keep_logs = keep_logs or (self._keep_logs_on_failure and not self.__created_clusters[name].create_complete)
+            delete_logs = delete_logs and (self._keep_logs_on_failure and self.__created_clusters[name].create_complete)
             try:
-                self.__created_clusters[name].delete(keep_logs=keep_logs)
-            except subprocess.CalledProcessError as e:
+                self.__created_clusters[name].delete(delete_logs=delete_logs)
+            except Exception as e:
                 logging.error(
-                    "Failed when deleting cluster %s with error %s. Retrying deletion without --keep-logs.", name, e
+                    "Failed when deleting cluster %s with error %s. Retrying deletion without deleting logs.", name, e
                 )
-                self.__created_clusters[name].delete(keep_logs=False)
+                self._destroy_cluster(name)
             del self.__created_clusters[name]
             logging.info("Cluster {0} deleted successfully".format(name))
         else:
             logging.warning("Couldn't find cluster with name {0}. Skipping deletion.".format(name))
 
-    def destroy_all_clusters(self, keep_logs=False):
+    @retry(stop_max_attempt_number=5, wait_fixed=5000, retry_on_exception=retry_if_subprocess_error)
+    def _destroy_cluster(self, name):
+        self.__created_clusters[name].delete(delete_logs=False)
+
+    def destroy_all_clusters(self, delete_logs):
         """Destroy all created clusters."""
         logging.debug("Destroying all clusters")
         for key in list(self.__created_clusters.keys()):
             try:
-                self.destroy_cluster(key, keep_logs)
+                self.destroy_cluster(key, delete_logs)
             except Exception as e:
                 logging.error("Failed when destroying cluster {0} with exception {1}.".format(key, e))