+++ tests/integration-tests/tests/resource_bucket/test_resource_bucket.py
@@ -9,7 +9,7 @@ from assertpy import assert_that
 @pytest.mark.schedulers(["slurm", "awsbatch"])
 @pytest.mark.oss(["alinux2"])
 @pytest.mark.usefixtures("os", "instance")
-def test_resource_bucket(region, scheduler, pcluster_config_reader, clusters_factory, s3_bucket_factory, test_datadir):
+def test_resource_bucket(region, scheduler, pcluster_config_reader, s3_bucket_factory, clusters_factory, test_datadir):
     # Bucket used to host cluster artifacts must have versioning enabled
     logging.info("Testing cluster creation/deletion behavior when specifying cluster_resource_bucket")
     bucket_name = s3_bucket_factory()
@@ -29,8 +29,7 @@ def test_resource_bucket(region, scheduler, pcluster_config_reader, clusters_fac
     updated_config_file = pcluster_config_reader(
         config_file="pcluster.config_{0}.yaml".format(scheduler), resource_bucket=update_bucket_name
     )
-    cluster.config_file = str(updated_config_file)
-    cluster.update()
+    cluster.update(str(updated_config_file))
     assert_that(cluster.cfn_parameters.get("ResourcesS3Bucket")).is_equal_to(bucket_name)
     assert_that(cluster.cfn_parameters.get("ArtifactS3RootDirectory")).is_equal_to(artifact_directory)